fetch_book_feeds/requirements.txt
feedparser==6.0.10
functions-framework==3.4.0
google-cloud-storage==2.10.0
requests==2.31.0